# bugfix(cs-perf): Temp Fix For CS Perf Issues